Hitting submit actually creates a task and saves to task list
So far, it doesn't actually store the tasks using localStorage, but it's a big step.
removed the demo tasks and created the logic needed to actually create and switch the tasks between lists.
Currently it just appends to the other list, so it does shift the positions of the tasks, not that I see that being a huge issue.
The Delete button also works as intended.